don't require scipy
