+++ src/diffusers/utils/dummy_torch_and_scipy_objects.py
@@ -4,36 +4,6 @@
 from ..utils import DummyObject, requires_backends
 
 
-class EulerAncestralDiscreteScheduler(metaclass=DummyObject):
-    _backends = ["torch", "scipy"]
-
-    def __init__(self, *args, **kwargs):
-        requires_backends(self, ["torch", "scipy"])
-
-    @classmethod
-    def from_config(cls, *args, **kwargs):
-        requires_backends(cls, ["torch", "scipy"])
-
-    @classmethod
-    def from_pretrained(cls, *args, **kwargs):
-        requires_backends(cls, ["torch", "scipy"])
-
-
-class EulerDiscreteScheduler(metaclass=DummyObject):
-    _backends = ["torch", "scipy"]
-
-    def __init__(self, *args, **kwargs):
-        requires_backends(self, ["torch", "scipy"])
-
-    @classmethod
-    def from_config(cls, *args, **kwargs):
-        requires_backends(cls, ["torch", "scipy"])
-
-    @classmethod
-    def from_pretrained(cls, *args, **kwargs):
-        requires_backends(cls, ["torch", "scipy"])
-
-
 class LMSDiscreteScheduler(metaclass=DummyObject):
     _backends = ["torch", "scipy"]
